Refactored all non visitor session related functions from PageActivity to MuistiActivity, so the functions will be available also on idle page
